Refactoring: Email Processing dir
